improved MaskBuilder
